Object construction and framework Restructuring
- Restructure framework to remove wordStorm
- Addition of framework to create objects
- minor changes to function levels of view
- restructured some data models
- restructure and change of how tweet Fetching will work
- Addition of Chart and Tweet implicits
- Adjustments to Sentiment Analysis

+++ dataScavenger/src/main/scala/com/flashboomlet/scavenger/polls/responseModels.scala
@@ -82,5 +82,5 @@ case class Point (
   */
 case class DataPoint (
   choice: String,
-  value: Float
+  value: Double
 )
